+++ statevar/add_noise_release.go
@@ -1,0 +1,10 @@
+//go:build !qa
+// +build !qa
+
+package statevar
+
+import vegapb "code.vegaprotocol.io/protos/vega"
+
+func AddNoise(kvb []*vegapb.KeyValueBundle) []*vegapb.KeyValueBundle {
+	return kvb
+}